refactor(agents): extract ActionEditorDialog from AgentActionsPanel
